+++ src/main/java/com/sudarshan/jpa/service/ContactService.java
@@ -10,11 +10,11 @@ public interface ContactService {
 	
 	public Contact add(ContactDTO added);
 	
-	public Contact deleteById(Long id) throws NotFoundException;
+	public Contact deleteById(int id) throws NotFoundException;
 	
 	public List<Contact> findAll();
 	
-	public Contact findById(Long id) throws NotFoundException;
+	public Contact findById(int id) throws NotFoundException;
 	
 	public Contact update(ContactDTO updated) throws NotFoundException;
